Move structs to central model location
